code review improvements

+++ core/src/main/java/io/questdb/cutlass/http/HttpResponseSink.java
@@ -480,7 +480,7 @@ public class HttpResponseSink implements Closeable, Mutable {
         @Override
         public int writeBytes(long srcAddr, int len) {
             assert len > 0;
-            len = (int) Math.min((long) len, buffer.getWriteNAvailable());
+            len = (int) Math.min(len, buffer.getWriteNAvailable());
             Vect.memcpy(buffer.getWriteAddress(len), srcAddr, len);
             buffer.onWrite(len);
             return len;
